week18 hw2 completed

+++ controllers/admin_controllers.js
@@ -4,6 +4,7 @@ const fs = require('fs');
 const User = db.User;
 const Prize = db.Prize;
 const Item = db.Item;
+const Faq = db.Faq;
 
 const admin_controller = {
   admin: async (req, res) => {
@@ -77,6 +78,10 @@ const admin_controller = {
       return res.redirect('/admin');
     }
     const { item_id, item_name , item_desc, item_proba } = req.fields;
+    if (!item_id || !item_name || !item_desc || !item_proba) {
+      req.flash('errorMessage', '獎項名、描述及機率不得為空');
+      return next()
+    }
     if(item_proba < 0) {
       req.flash('errorMessage', '不得輸入負數');
       return next()
@@ -147,13 +152,19 @@ const admin_controller = {
     if (!userId) {
       return res.redirect('/admin');
     }
-    const { item_id } = req.params;
+    let { item_id } = req.params;
+    item_id = parseInt(item_id);
 
-    await Prize.destroy({
-      where: {
-        id:item_id
-      }
-    })
+    if(item_id) {
+      await Prize.destroy({
+        where: {
+          id:item_id
+        }
+      })
+    } else {
+      req.flash('errorMessage', '刪除失敗');
+      return next()
+    }
 
     return next();
   },
@@ -164,6 +175,10 @@ const admin_controller = {
       return res.redirect('/admin');
     }
     const { item_name , item_desc } = req.fields;
+    if (!item_name || !item_desc) {
+      req.flash('errorMessage', '獎項名及描述不得為空');
+      return next()
+    }
     const { item_image } = req.files
 
     if (!item_image.size) {
@@ -240,6 +255,10 @@ const admin_controller = {
     }
 
     const { product_id, product_name, product_price, product_quantity} = req.fields;
+    if (!product_name || !product_price || !product_quantity) {
+      req.flash('errorMessage', '品名、價格及數量不得為空');
+      return next()
+    }
     const { product_image } = req.files;
 
     if (!product_image.size) {
@@ -310,6 +329,10 @@ const admin_controller = {
       return res.redirect('/admin');
     }
     const { product_name, product_price, product_quantity} = req.fields;
+    if (!product_name || !product_price || !product_quantity) {
+      req.flash('errorMessage', '品名、價格及數量不得為空');
+      return next()
+    }
     const { product_image } = req.files;
 
     if (!product_image.size) {
@@ -343,6 +366,81 @@ const admin_controller = {
       item_quantity:product_quantity,
       item_image:product_image.name
     })
+    return next()
+  },
+
+  adminFaq: async (req, res) => {
+    const { userId } = req.session;
+    if (!userId) {
+      return res.redirect('/admin');
+    }
+    const faqs = await Faq.findAll({
+      order: [['faq_order', 'ASC']]
+    });
+    
+    res.render('admin_faq', { faqs:faqs });
+  },
+
+  handleUpdateFaq: async (req, res, next) => {
+    const { userId } = req.session;
+    if (!userId) {
+      return res.redirect('/admin');
+    }
+    const { faq_id, faq_title, faq_content, faq_order } = req.body;
+
+    if (!faq_title || !faq_content || !faq_order) {
+      req.flash('errorMessage', '標題、內容及順序不得為空');
+      return next()
+    }
+
+    Faq.update({
+      faq_title:faq_title,
+      faq_content:faq_content,
+      faq_order:faq_order
+    },{
+      where: {
+        id:faq_id
+      }
+    })
+
+    return next()
+  },
+
+  handleDeleteFaq: async (req, res, next) => {
+    const { userId } = req.session;
+    if (!userId) {
+      return res.redirect('/admin');
+    }
+    let { faq_id } = req.params;
+    faq_id = parseInt(faq_id);
+    if(faq_id) {
+      Faq.destroy({
+        where:{
+          id:faq_id
+        }
+      });
+    } else {
+      req.flash('errorMessage', '刪除失敗');
+      return next()
+    }
+    return next();
+  },
+  handleCreateFaq: async (req, res, next) => {
+    const { userId } = req.session;
+    if (!userId) {
+      return res.redirect('/admin');
+    }
+    const { faq_title, faq_content, faq_order } = req.body;
+    if (!faq_title || !faq_content || !faq_order) {
+      req.flash('errorMessage', '標題、內容及順序不得為空');
+      return next()
+    }
+    Faq.create({
+      faq_title:faq_title,
+      faq_content:faq_content,
+      faq_order:parseInt(faq_order)
+    })
+
     return next()
   }
 };
